# backend/src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blockchain Identity System API",
    "description": "RESTful API for managing and verifying blockchain-based digital identities",
    "version": "1.0.0",
    "contact": {
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.identity.example.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/system/stats": {
      "get": {
        "tags": ["System"],
        "summary": "System statistics",
        "description": "Get system-wide statistics",
        "operationId": "getSystemStats",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "System statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "totalIdentitiesMinted": {
                          "type": "string",
                          "example": "1000"
                        },
                        "totalModerationCases": {
                          "type": "string",
                          "example": "50"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2023-03-15T14:30:00Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/system/apikeys": {
      "post": {
        "tags": ["System"],
        "summary": "Generate API key",
        "description": "Generate a new API key for third-party integration",
        "operationId": "generateApiKey",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["clientId", "clientName"],
                "properties": {
                  "clientId": {
                    "type": "string",
                    "description": "Client identifier",
                    "example": "client-123"
                  },
                  "clientName": {
                    "type": "string",
                    "description": "Client name",
                    "example": "Test Client"
                  },
                  "permissions": {
                    "type": "array",
                    "description": "Permission strings",
                    "items": {
                      "type": "string"
                    },
                    "example": ["identity.read", "reputation.read"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "API key generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "apiKey": {
                          "type": "string",
                          "example": "f8d132e7c5884a6d89f55f1e77248e1f"
                        },
                        "clientId": {
                          "type": "string",
                          "example": "client-123"
                        },
                        "clientName": {
                          "type": "string",
                          "example": "Test Client"
                        },
                        "permissions": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "example": ["identity.read", "reputation.read"]
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2023-03-15T14:30:00Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["System"],
        "summary": "Deactivate API key",
        "description": "Deactivate an existing API key",
        "operationId": "deactivateApiKey",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["apiKey"],
                "properties": {
                  "apiKey": {
                    "type": "string",
                    "description": "API key to deactivate",
                    "example": "f8d132e7c5884a6d89f55f1e77248e1f"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "API key deactivated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "API key deactivated successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/system/apikeys/{clientId}": {
      "get": {
        "tags": ["System"],
        "summary": "Get client API keys",
        "description": "Get all API keys for a specific client",
        "operationId": "getClientApiKeys",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "description": "Client identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "API keys retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "clientId": {
                          "type": "string",
                          "example": "client-123"
                        },
                        "keys": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "key": {
                                "type": "string",
                                "example": "f8d132e7..."
                              },
                              "active": {
                                "type": "boolean",
                                "example": true
                              },
                              "createdAt": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2023-03-15T14:30:00Z"
                              },
                              "permissions": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                },
                                "example": ["identity.read", "reputation.read"]
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/system/devkey": {
      "get": {
        "tags": ["System"],
        "summary": "Get development API key",
        "description": "Generate a new API key for development purposes (only available in development environment)",
        "operationId": "getDevKey",
        "responses": {
          "200": {
            "description": "API key generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Development API key generated"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "apiKey": {
                          "type": "string",
                          "example": "f8d132e7c5884a6d89f55f1e77248e1f"
                        },
                        "note": {
                          "type": "string",
                          "example": "This key is for development purposes only and will be lost when the server restarts"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden in production",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Error message",
            "example": "Invalid signature"
          },
          "code": {
            "type": "string",
            "description": "Error code (optional)",
            "example": "INVALID_SIGNATURE"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for gateway access"
      }
    }
  }
}